Fix panics triggered by validating nested unknown values
The original issue motivating this patch:

        https://github.com/pulumi/pulumi-aws/issues/2818

The context here is that AWS provider has provider schema defined as:

        {assume_role: [{duration: durationType}]}

When a Pulumi explicit provider has unknowns under assumeRole, it is pasing this datum in:

        {assume_role: [unknown]}

Prior to the fix this used to cause a panic. Data.ValueAtPath is called to validate the duration
attribute, and attempts to drill down the path of assume_role[0].duration, fails to notice the
unknown, and passes an uninitialied value into the duration validator, which then panics.

It is possible to have an alternative fix where Data.ValueAtPath is not being called at all when
drilling down an Unknown at the level of fwserver.BlockValidate or higher up the stack.

Still it seems resonable to make Data.ValueAtPath itself more robust by this change. Sending
uninitialized value into the validators is not a good idea as they may not be prepared to handle
such values.